Functions package, that implements php-ews, built to manage shared contacts
Problem:
There is a need to access, update, create, and delete the contact
information stored on an exchange 2010 (service pack 2) server.  More
specifically, this information needs to be accessed from an application
using Exchange Web Services and PHP. The contacts that need to be
manipulated are being shared with a ‘dummy’ exchange account, which is
given read/write access. The solution only needs to be able to
manipulate  “stock” contact fields.

Exchange Interface Package:
The exchange interface provides a simple functions package that allows
application developers to manipulate exchange contacts. The exchange
interface is built upon the php-ews package
(https://github.com/jamesiarmes/php-ews) and shares all dependencies.